Cache declared queues and exchanges (#11)
Declared queues, exchanges and consumes are now cached and will not be attempted to be re-declared on subsequent calls to consume/publish interfaces. The caching is based on resource names. A re-consume of the same queue name will (as long as the combination of queue+exchange+routing key isn't the same) be let through, resources that are not yet declared and declared and bindings are set up. However, since the consume of the queue already exists basic_consume will not be called again (so it is not possible to change callback this way), rather the existing consume callback will receive a ConsumeOK with the previous consumer tag as an indication that the consume was successful but that the consumer tag is the same.